+++ src/NServiceBus.TransactionalSession.Tests/Fakes/FakeDispatcher.cs
@@ -1,0 +1,18 @@
+﻿namespace NServiceBus.TransactionalSession.Tests.Fakes;
+
+using System.Collections.Generic;
+using System.Threading;
+using System.Threading.Tasks;
+using Transport;
+
+class FakeDispatcher : IMessageDispatcher
+{
+    public List<(TransportOperations outgoingMessages, TransportTransaction transaction)> Dispatched = new();
+
+    public Task Dispatch(TransportOperations outgoingMessages, TransportTransaction transaction,
+        CancellationToken cancellationToken = new CancellationToken())
+    {
+        Dispatched.Add((outgoingMessages, transaction));
+        return Task.CompletedTask;
+    }
+}
